refactor(core): move TableFeature to metadata.features package
