+++ core/src/main/kotlin/com/kakao/actionbase/core/metadata/features/TableFeature.kt
@@ -1,0 +1,9 @@
+package com.kakao.actionbase.core.metadata.features
+
+enum class TableFeature {
+    /**
+     * INSERT merges into the existing row: an omitted field keeps its current value instead of
+     * being cleared to UNSET (snapshot).
+     */
+    INSERT_MERGE,
+}
